Replace Quantity Text input with Number Input
Changes the Quantity input into a Number input, adding an Increase and Decrease arrow to the box, also make it where the number keypad opens by default on mobile phone and tablets.  (This does not stop users from adding text to the input, just makes it look more professional, especially on mobile devices).